requirements.txt
asgiref==3.9.1
distlib==0.4.0
Django==5.2.4
django-cors-headers==4.7.0
djangorestframework==3.16.1
filelock==3.19.1
mysql==0.0.3
mysqlclient==2.2.7
platformdirs==4.4.0
sqlparse==0.5.3
tzdata==2025.2
virtualenv==20.34.0
